a) Enabled the new test framework for executing runtime tests. b) Modified testsiute.xml to comment out all tests that were listed as part of runtimets/ignore.txt c) Removed RuntimeTest.java, runtimets/ignore.txt and runtimets/only.txt as they are now redundant.
Note:- In the event that you want to add a test case under the runtime tests, please follow instructions 
corresponding to the new test framework. 
If you want to run a subset of  the runtime testsuite (a functionality supported through use of only.txt), you 
currently do not have a way to do so with the new test framework.  Other test suites (metadata, optimizer etc) 
continue to run via the old test framework. 


git-svn-id: http://asterixdb.googlecode.com/svn/branches/asterix_stabilization@851 eaa15691-b419-025a-1212-ee371bd00084